modifieed next-app and updated readme
