+++ next-app/pages/index.js
@@ -1,23 +1,25 @@
-import ButtomP from "../components/submitForm"
+import Forms from "../components/submitForm"
 import Table from "../components/table"
-import DateTime from "../components/dateTime" 
+import DateTime from "../components/dateTime"
+import AdminPanel from "../components/Admin"
 
 export default function Home() {
   return (
-<>
-    
-    <DateTime/>
-    <ButtomP/>
-    <Table />
-</>
+    <>
+
+      <DateTime />
+      <Forms />
+      <AdminPanel />
+      <Table />
+    </>
   )
 }
 
 
 export async function getServerSideProps(ctx) {
   // Fetch data from external API
- 
+
   console.log(ctx.query)
   // Pass data to the page via props
-  return { props: {  } }
+  return { props: {} }
 }
