+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reactstrap-react-lib",
-  "version": "1.2.6",
+  "version": "1.2.7",
   "description": "This is a lib for react and next",
   "main": "lib/index.js",
   "types": "lib",
